refine
- fix license headers
- put shared app component into separate library
- document app
