+++ CMakeLists.txt
@@ -144,6 +144,9 @@ if (PYCLINGODL_ENABLE STREQUAL "pip")
     elseif (APPLE)
         target_link_options(libclingo INTERFACE "-undefined" "dynamic_lookup")
     endif()
+    if (CLINGODL_BUILD_TESTS)
+        add_subdirectory(app/lib)
+    endif()
     add_subdirectory(libclingo-dl)
     add_subdirectory(libpyclingo-dl)
 else()
