+++ LICENSE.md
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2018 Roland Kaminski, Philipp Wanko, Max Ostrowski
+Copyright (c) Roland Kaminski, Philipp Wanko, and Max Ostrowski
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
