+++ app/CMakeLists.txt
@@ -1,14 +1,16 @@
-# [[[source: .
+# [[[source: src
 set(ide_source_group "Source Files")
 set(source-group
-    "${CMAKE_CURRENT_SOURCE_DIR}/main.cc")
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/main.cc")
 source_group("${ide_source_group}" FILES ${source-group})
 set(source
     ${source-group})
 # ]]]
 
+add_subdirectory(lib)
+
 add_executable(clingo-dl ${source})
-target_link_libraries(clingo-dl PRIVATE libclingo libclingo-dl)
+target_link_libraries(clingo-dl PRIVATE libclingo libclingo-dl-app)
 set_target_properties(clingo-dl PROPERTIES FOLDER exe)
 
 install(TARGETS clingo-dl
