+++ libclingo-dl/CMakeLists.txt
@@ -4,7 +4,6 @@ set(header-group
     "${CMAKE_CURRENT_SOURCE_DIR}/clingo-dl.h")
 source_group("${ide_header_group}" FILES ${header-group})
 set(header-group-clingo-dl
-    "${CMAKE_CURRENT_SOURCE_DIR}/clingo-dl/app.hh"
     "${CMAKE_CURRENT_SOURCE_DIR}/clingo-dl/config.hh"
     "${CMAKE_CURRENT_SOURCE_DIR}/clingo-dl/graph.hh"
     "${CMAKE_CURRENT_SOURCE_DIR}/clingo-dl/parsing.hh"
