+++ libclingo-dl/src/theory.cc
@@ -1,6 +1,6 @@
 // {{{ MIT License
 //
-// Copyright 2018 Roland Kaminski, Philipp Wanko, Max Ostrowski
+// Copyright Roland Kaminski, Philipp Wanko, and Max Ostrowski
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to
